Extract possibly sharable KtorHttpServer.
